+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- `--pileup` option to ouptu full raw base counts rather than BED methyl.
+### Changed
+- `-c` no longer synonym to `--cpg`.
+### Fixed
+- Links in README.
+
 ## [v0.6.3]
 ### Changed
 - Bumped htslib version to version 1.16 for fixes to MM tag parsing/validation.